♻️ Fold EventCalendarCard into the scheduled-event datetime components
Move the CalendarCard primitives to components/ and EventCalendarCard
next to EventDate/EventTimeRange, sharing their SchedulingFields
semantics: floating times read as UTC wall time, fixed instants show in
the viewer's zone. Also gate it behind useHydrated like the other
datetime components.
